OLD_FILES/lens_old/multi_model_controller/ai_query_service/requirements.txt
fastapi
uvicorn
psycopg2
openai
python-dotenv

